fix bugs in cpu calculation related to overflow

+++ src/tools.cpp
@@ -23,35 +23,35 @@ namespace mui
 
         if (fp.is_open())
         {
-            int i = 0;      // CPU counter
+            int i = 0;
             std::string line;
-            while (std::getline(fp, line) && i < 1) // End of file or 4 cpus founded
+            while (std::getline(fp, line) && i < 1)
             {
                 istringstream ss(line);
                 string field0;
                 ss >> field0;
 
-                char cpu_id[5]; // CPU to search
-                sprintf(cpu_id, "cpu%d", i);  // Get CPU ID
+                char cpu_id[5];
+                sprintf(cpu_id, "cpu%d", i);
 
                 if (field0.find(cpu_id) != std::string::npos)
                 {
-                    int times[7];
+                    uint64_t times[7];
                     for (auto& x : times)
                         ss >> x;
 
-                    int work_time = times[0] + times[1] + times[2];
-                    int total_time = work_time + times[3] + times[4] + times[5] + times[6];
+                    uint64_t work_time = times[0] + times[1] + times[2];
+                    uint64_t total_time = work_time + times[3] + times[4] + times[5] + times[6];
 
                     // Update CPU Usage
-                    m_cpu_usage[i] = ((float)work_time - m_work_cpu_last_time[i]) /
-                                     ((float)total_time - m_total_cpu_last_time[i]) * 100;
+                    m_cpu_usage[i] = ((double)work_time - m_work_cpu_last_time[i]) /
+                                     ((double)total_time - m_total_cpu_last_time[i]) * 100.;
 
                     // Update last values
-                    m_work_cpu_last_time[i] = (float)work_time;
-                    m_total_cpu_last_time[i] = (float)total_time;
+                    m_work_cpu_last_time[i] = (double)work_time;
+                    m_total_cpu_last_time[i] = (double)total_time;
 
-                    i++; // Next CPU
+                    i++;
                 }
             }
